Changed controller order for GUI interface instead of text
Changed controller order for GUI interface instead of text.
Removed the controller list from the Configuration->Input.
Added the controller list to the pin.html as part of each named device.
Using alt_controller_count_temp and alt_controller_enable_temp in the pin.html which is really the settings from the Config->Input for the alternate controller number.