Add share controller mode. (strict, share, both)
The controller mode did not have a setting before. Now it does.
Before the only controller mode was the strict mode. Now the strict, share, and both modes are available.

Added the ability to have shared host controllers between 2 or more client controllers. This means that 2 or more people can control the same controller on the host side.
//////////////////////////////////
On the configuration->input web page, there are now the following options:
Enable alternative controller numbering mode (Existed in original pull request)
---Number of Controllers to preallocate (Existed in original pull request)
--- Alternate Controller Mode (new)
------ Mode 1: Strict (was really the default and not an option in the original pull request)
------ Mode 2: Shared ( new )
------ Mode 3 : Both (new) – Has ability to switch between both modes via SHIFT+CTRL+ALT+I keypress
//////////////////////////////////
On the PIN web options per named device/client, there are now the following options (if the alternative controller numbering mode is enabled)
--- Alternate Controller Strict Mode List (Existed in original pull request)
--- Alternate Controller Shared Mode List (New)
--- Alternative Controller Jitter Joystick Correction List (New) – Typically needed when sharing controllers so that an “idle” controller does not seize control.
--- Alternative Controller Swap Joysticks List ( New )